refactored DocEmbeddings as EmbeddingSearcher
